Add user options for Game and Author
