+++ main.rb
@@ -1,44 +1,11 @@
-def run
-  puts 'Please choose an option by entering a number:'
-  puts '1 - List all books'
-  puts '2 - List all music albums'
-  puts '3 - List all games'
-  puts '4 - List all genres'
-  puts '5 - List all labels'
-  puts '6 - List all authors'
-  puts '7 - Add a book'
-  puts '8 - Add a music album'
-  puts '9 - Add a game'
-  puts '10 - Exit'
-end
-
-OPTIONS = {
-  1 => :list_books,
-  2 => :list_music_albums,
-  3 => :list_games,
-  4 => :list_genres,
-  5 => :list_labels,
-  6 => :list_authors,
-  7 => :add_book,
-  8 => :add_music_album,
-  9 => :add_game,
-  10 => :exit
-}.freeze
-
-def option(option)
-  input_method = OPTIONS[option]
-  if input_method
-    send(input_method)
-  else
-    puts 'Invalid input, please input correct number!'
-  end
-end
+require_relative 'app'
 
 def main
+  app = App.new
   loop do
-    run
+    app.run
     option = gets.chomp.to_i
-    option(option)
+    app.option(option)
   end
 end
